Fix scoping rules
When we were doing scoping + generation at the same time, we always had
the local scope at hand when we were generating code. By adding a scope
pass before generating, the only scope that the gen pass was inheriting
was the global scope for a given descriptor (i.e., the scope for the
entire proto file). This meant that scope searches for nested types
(either enums or messages) would fail, as they didn't have the
appropriate prefixes in place.

Now, we add a mapping of identifier to scope, and at generate time we
look up the appropriate local scope for the type we're generating, so
scoping should always be successful.