Setup standard pony project release notes handling
This adds everything related to building up release notes that isn't
the rotation of release notes after a release (that's handled by the
release-bot-action which will be installed later).

For this to work, a new secret needs to be added to this repository
under the name API_TOKEN. API_TOKEN should be a GitHub API token that
allows public repo access. You should add the token under your ergl
account.

Note, the release-notes-bot will skip adding release notes for any PR
that doesn't have one of the changelog - * labels.