Add tests that check that code works with the latest ponyc (#8)
This adds a test that runs every evening against the most recent
ponyc build to make sure that everything still works with the latest
ponyc changes.

This particular configuration uses GitHub notifications to trigger
the run. For this to work, you'll need to invite the user ponylang-main
to be a collaborator on the project and that account will have to accept.

Additionally, a PR to send a message to this repo when the nightly builders
are updated will need to be added to the `ponylang/shared-docker` repo.

This workflow could also be run on a schedule but that, while less work
to set up, is also less precise.